# switch to PrecompileTools.jl